# Conflicts:
#	src/survive_vive.c

+++ src/survive_config.h
@@ -35,7 +35,7 @@ typedef struct config_group {
 //extern config_group global_config_values;
 //extern config_group lh_config[2]; //lighthouse configs
 
-void init_config_group(config_group *cg, uint16_t count);
+void init_config_group(config_group *cg, uint8_t count);
 void destroy_config_group(config_group* cg);
 
 //void config_init();
@@ -52,7 +52,7 @@ const uint32_t config_set_uint32(config_group *cg, const char *tag, const uint32
 const char* config_set_str(config_group *cg, const char *tag, const char* value);
 
 FLT config_read_float(config_group *cg, const char *tag, const FLT def);
-uint16_t config_read_float_array(config_group *cg, const char *tag, const FLT** values, const FLT* def, uint16_t count);
+uint16_t config_read_float_array(config_group *cg, const char *tag, const FLT** values, const FLT* def, uint8_t count);
 uint32_t config_read_uint32(config_group *cg, const char *tag, const uint32_t def);
 const char* config_read_str(config_group *cg, const char *tag, const char *def);
